force python3-Pygments installation
Pygments is available in the 5.15 kde sdk but not the plaform runtime so
it's not being installed during build-time and it's missing during
runtime.
This workaround uses the `pip install` option `--ignore-installed` to
force-install the module into `$FLATPAK_DEST` without ending with an
error after pip try and fail to make changes to the read-only `/usr`.