Build our own QtWebEngine lib
I'm not really thrilled about this but it's better than shipping a
browser with multiple security vulnerabilities.
I was hoping to get a more prompt response from the QtWebEngine
maintainer in flathub/io.qt.qtwebengine.BaseApp#15 but this doesn't
seem to be happening.

I should stress that this is a stop gap until the QtWebEngine base app
will be updated. I have no desire to keep it like this as I'm not able
nor I'm willing to build Chromium locally.

For quickly building locally and testing packaging changes just revert or
uncomment the base app related properties and comment out the qt5-webengine
module.