Drop fedc.json manifests of make-only depends
Simplify auto-updating by merging the separate fedc.json manifests of
make-only dependencies into the respected module's manifests.
They were added to avoid unneeded rebuilds when we were also building
PyQt and userscripts' dependencies, but now we're building the app
pretty quickly, so this is less of a concern, and f-e-d-c is gaining
selective updating capability, which we might be able to take advantage of.